Update README.md
It’s the maximum amount of share capital that a Company is allowed to issue to its shareholders as per its constitutional documents. Shares are defined as the financial instruments that form units of the overall capital and it is used to raise funds from the public. It is used to limit the ability of Directors to allot new shares which may have consequences over the control of the Company. It’s used to prevent any shift in the profit distribution balance. Often, the amount provided as Authorized Capital is not fully used & a small % to be kept as a safety buffer to raise additional capital when you need to increase.